+++ client/src/components/CheckoutForm.test.js
@@ -4,6 +4,8 @@ import CheckoutForm from "./CheckoutForm";
 
 // Write up the two tests here and make sure they are testing what the title shows
 
-test("form header renders", () => {});
+test("form header renders", () => {
+    render(<CheckoutForm/>)
+});
 
 test("form shows success message on submit with form details", () => {});
