+++ client/src/components/PlantList.js
@@ -3,7 +3,27 @@ import axios from "axios";
 
 export default class PlantList extends Component {
   // add state with a property called "plants" - initialize as an empty array
+  constructor(){
+    super()
+    this.state = {
+      plants: []
+    }   
+  }
 
+componentDidMount() {
+  axios
+    .get('http://localhost:3333/plants')
+    .then(res => {
+      console.log(res)
+      this.setState({
+        plants: res.data.plantsData,
+      })
+    })
+    .catch(err => {
+      debugger
+      console.log(err)
+    });
+}
   // when the component mounts:
   //   - fetch data from the server endpoint - http://localhost:3333/plants
   //   - set the returned plants array to this.state.plants
